+++ email version/css/style.css
@@ -1,6 +1,3 @@
-
-
-
 footer{
     position:fixed;
     left:0px;
@@ -11,7 +8,7 @@ footer{
     overflow: hidden; 
  }
 
-#ulfooter{ 
+ ul { 
     list-style: none;
     position: relative;
     float: left; 
@@ -19,10 +16,10 @@ footer{
     left: 45%;
     margin-top: 20px;
  }
- #lets{ 
+ li { 
     position: relative;
     float: left;
-    display: inline;
+    display: block;
     right: 50%; 
 }
 a:link {
