+++ email version/images/.DS_Store
