+++ email version/index.html
@@ -1,0 +1,127 @@
+    <!DOCTYPE html>
+    <html lang="en">
+
+    <head>
+        <!-- Required meta tags -->
+        <meta charset="utf-8">
+        <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
+
+        <!-- Bootstrap CSS -->
+        <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/4.0.0/css/bootstrap.min.css" integrity="sha384-Gn5384xqQ1aoWXA+058RXPxPg6fy4IWvTNh0E263XmFcJlSAwiGgFAW/dAiS6JXm"
+            crossorigin="anonymous">
+
+
+        <!-- firebase linking script -->
+
+        <link rel="stylesheet" href="assets/css/style.css" type="text/css">
+
+        <title>Cryptobot</title>
+    </head>
+
+    <body>
+
+        <nav class="navbar navbar-expand-lg navbar-light bg-dark mb-4" style="padding-top: 0%; padding-bottom: 0%; margin-top: 0%; margin-bottom: 0%;">
+            <a class="navbar-brand" href="#">
+                <img src="assets/images/logof.png" width="60px" style="padding-top: 0%; padding-bottom: 0%;" alt="">
+                <img src="assets/images/logowofb.png" width="120px;" style="padding-top: 0%; padding-bottom: 0%; margin-top: 0%; margin-bottom: 0%;"
+                    alt="">
+            </a>
+            <button class="navbar-toggler" type="button" data-toggle="collapse" data-target="#navbarSupportedContent" aria-controls="navbarSupportedContent"
+                aria-expanded="false" aria-label="Toggle navigation">
+                <span class="navbar-toggler-icon"></span>
+            </button>
+
+            <div class="collapse navbar-collapse" id="navbarSupportedContent">
+                <ul class="navbar-nav mr-auto">
+                    <li class="nav-item active">
+                        <a class="nav-link" href="#" style="display: none;">History
+                            <span class="sr-only">(current)</span>
+                        </a>
+                    </li>
+                </ul>
+            </div>
+        </nav>
+
+
+
+        <div class="row">
+            <div class="col-8">
+                <h1 id="bigOne"></h1>
+            </div>
+            <div class="col-4">
+                <div class="container px-4 py-4" style="border: darkgrey; border-style: solid; border-width: 2px; border-radius: 15px; width: 350px;">
+
+                    <div class="container">
+                        <div class="form-group">
+                            <form class="px-4 py-3">
+                                <div class="form-group">
+                                    <label for="exampleDropdownFormEmail1">Email address</label>
+                                    <input type="email" class="form-control" id="txtEmail" placeholder="email@example.com">
+                                </div>
+                                <div class="form-group">
+                                    <label for="exampleDropdownFormPassword1">Password</label>
+                                    <input type="password" class="form-control" id="txtPassword" placeholder="Password">
+                                </div>
+                                <div class="form-check my-2">
+                                    <input type="checkbox" class="form-check-input" id="dropdownCheck">
+                                    <label class="form-check-label" for="dropdownCheck">
+                                        Remember me
+                                    </label>
+                                </div>
+                                <button class="btn btn-success my-2 mx-2 px-3 py-2" id="btnLogin">Sign in</button>
+                                <button  class="btn btn-dark my-2 mx-2 px-3 py-2" id="btnSignUp">Sign Up</button>
+                                <button  class="btn btn-dark my-2 mx-2 px-3 py-2 invisible" id="btnLogOut">Log Out</button>
+
+                            </form>
+                            <div class="dropdown-divider"></div>
+                            <a class="dropdown-item" href="#">Forgot password?</a>
+                        </div>
+
+                    </div>
+
+                </div>
+            </div>
+
+        </div>
+
+
+
+        <footer style="background-color: darkgrey;  ">
+
+            <ul>
+                <li>
+                    <a href="#">Contact us &nbsp;&nbsp;</a>
+                </li>
+                <li>
+                    <a href="#">Support &nbsp;&nbsp;</a>
+                </li>
+                <li>
+                    <a href="#">Our Team</a>
+                </li>
+            </ul>
+
+        </footer>
+
+
+
+        <script src="https://www.gstatic.com/firebasejs/5.0.4/firebase.js"></script>
+
+
+
+        <!-- Optional JavaScript -->
+        <!-- jQuery first, then Popper.js, then Bootstrap JS -->
+        <script src="https://code.jquery.com/jquery-3.2.1.slim.min.js" integrity="sha384-KJ3o2DKtIkvYIK3UENzmM7KCkRr/rE9/Qpg6aAZGJwFDMVNA/GpGFF93hXpG5KkN"
+            crossorigin="anonymous"></script>
+        <script src="https://cdnjs.cloudflare.com/ajax/libs/popper.js/1.12.9/umd/popper.min.js" integrity="sha384-ApNbgh9B+Y1QKtv3Rn7W3mgPxhU9K/ScQsAP7hUibX39j7fakFPskvXusvfa0b4Q"
+            crossorigin="anonymous"></script>
+        <script src="https://maxcdn.bootstrapcdn.com/bootstrap/4.0.0/js/bootstrap.min.js" integrity="sha384-JZR6Spejh4U02d8jOt6vLEHfe/JQGiRRSQQxSfFWpi1MquVdAyjUar5+76PVCmYl"
+            crossorigin="anonymous"></script>
+
+        <script src="assets/javascript/auth.js"></script>
+
+
+
+
+    </body>
+
+    </html>
