+++ email version/javascript/bot.js
