+++ email version/javascript/cryptobot.html
@@ -48,6 +48,7 @@
                 </li>
             </ul>
            
+            
             <form class="form-inline my-2 my-lg-0">
                 <button class="btn btn-light my-2 my-sm-0 visible" id="btnLogOut">Log Out</button>
             </form>
