+++ email version/javascript/out.js
@@ -4,6 +4,7 @@
     const btnLogOut = document.getElementById('btnLogOut');
 
     btnLogOut.addEventListener('click', e => {
+       
         e.preventDefault();
 
         firebase.auth().signOut();
