+++ profile.html
@@ -1,0 +1,178 @@
+<!DOCTYPE html>
+<html lang="en">
+
+<head>
+    <meta charset="utf-8">
+    <meta http-equiv="X-UA-Compatible" content="IE=edge">
+    <meta name="viewport" content="width=device-width, initial-scale=1">
+    <title>Firebase - Profile</title>
+    <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.1.0/css/bootstrap.min.css" integrity="sha384-9gVQ4dYFwwWSjIDZnLEWnxCjeSWFphJiwGPXr1jddIhOegiu1FwO5qRGvFXOdJZ4"
+    crossorigin="anonymous">
+    <link rel="stylesheet" href="assets/css/style.css" type="text/css">
+
+    <!-- Bootstrap -->
+
+    
+</head>
+
+<body>
+
+        <nav class="navbar navbar-expand-lg navbar-light bg-dark">
+                <a class="navbar-brand" style="width: 600px;" href="index.html">
+                    <img src="assets/images/logof.png" width="10%">
+                    
+                    <img src="assets/images/logowofb.png" alt="" width="13%">
+                </a>
+                <button class="navbar-toggler" type="button" data-toggle="collapse" data-target="#navbarSupportedContent" aria-controls="navbarSupportedContent"
+                    aria-expanded="false" aria-label="Toggle navigation">
+                    <span class="navbar-toggler-icon"></span>
+                </button>
+        
+                <div class="collapse navbar-collapse" id="navbarSupportedContent">
+                    <ul class="navbar-nav mr-auto">
+                        <li class="nav-item active">
+                            <!-- <a class="nav-link" href="#">Home
+                                <span class="sr-only">(current)</span>
+                            </a> -->
+                        </li>
+                        <!--  profile <li class="nav-item">
+                            <a class="nav-link" href="create.html">Link</a>
+                        </li> -->
+                        <li class="nav-item">
+                            <a class="nav-link invisible" href="create.html">Link</a>
+                        </li>
+        
+                        <!-- <li class="nav-item dropdown">
+                            <a class="nav-link dropdown-toggle" href="#" id="navbarDropdown" role="button" data-toggle="dropdown" aria-haspopup="true"
+                                aria-expanded="false">
+                                Dropdown
+                            </a>
+                            <div class="dropdown-menu" aria-labelledby="navbarDropdown">
+                                <a class="dropdown-item" href="#">Action</a>
+                                <a class="dropdown-item" href="#">Another action</a>
+                                <div class="dropdown-divider"></div>
+                                <a class="dropdown-item" href="#">Something else here</a>
+                            </div>
+                        </li>
+                        <li class="nav-item">
+                            <a class="nav-link disabled" href="#">Disabled</a>
+                        </li> -->
+                    </ul>
+        
+                    <a class="btn btn-success" href="login.html" role="button">Sign In</a>
+                    <button id="logout" class="btn btn-light ml-2">Logout</button>
+        
+        
+                </div>
+            </nav>
+    </div>
+ 
+ 
+    <div class="row">
+        <div class="input-group col-md-4 mx-2 my-4">
+            <div class="input-group-append">
+                <span class="input-group-text" >$</span>
+            </div>
+            <input type="text" class="form-control" aria-label="Amount (to the nearest dollar)" placeholder="Amount to Trade">
+        </div>
+        <div class="input-group col-md-4 mx-2 my-4">
+            <input type="text" class="form-control" aria-label="Text input with dropdown button" placeholder="Amount or percentage">
+            <div class="input-group-append">
+                <button class="btn btn-secondary  dropdown-toggle" type="button" data-toggle="dropdown" aria-haspopup="true" aria-expanded="false">Stop Loss</button>
+                <div class="dropdown-menu ">
+                    <a class="dropdown-item" href="#">$</a>
+                    <a class="dropdown-item" href="#">%</a>
+                </div>
+            </div>
+        </div>
+
+        <button type="button" class="btn btn-success col-md-2 mx-4 my-4">Start</button>
+
+    </div>
+
+ 
+
+    <footer style="background-color: darkgrey;  ">
+
+         
+        <ul id="ulfooter">
+                <li id="lets">
+                    <a href="#">Contact us &nbsp;&nbsp;</a>
+                </li>
+                <li id="lets">
+                    <a href="#">Support &nbsp;&nbsp;</a>
+                </li>
+                <li id="lets">
+                    <a href="#">Our Team</a>
+                </li>
+            </ul>
+        
+    
+        </footer>
+ 
+ 
+    <script src="https://ajax.googleapis.com/ajax/libs/jquery/1.11.2/jquery.min.js"></script>
+    <script src="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.2/js/bootstrap.min.js"></script>
+    <script src="https://www.gstatic.com/firebasejs/3.3.0/firebase.js"></script>
+    <script src="https://www.gstatic.com/firebasejs/5.0.4/firebase.js"></script>
+
+    <script>
+        // firebase config
+
+
+
+  // Initialize Firebase
+  var config = {
+    apiKey: "AIzaSyB_gHsrrFKG_qgTUzD6w7Ybuqbfw0ayH8I",
+    authDomain: "cryptobot-firebase.firebaseapp.com",
+    databaseURL: "https://cryptobot-firebase.firebaseio.com",
+    projectId: "cryptobot-firebase",
+    storageBucket: "cryptobot-firebase.appspot.com",
+    messagingSenderId: "447326953953"
+  };
+  firebase.initializeApp(config);
+
+
+        /////////////////////////////////////
+
+
+        /**********************\
+         * Check login status *
+        \**********************/
+
+        firebase.auth().onAuthStateChanged(function (user) {
+            if (user) {
+
+                console.log(user);
+                document.getElementById('profile').innerHTML = JSON.stringify(user, null, 2);
+                document.getElementById('logout').onclick = function () {
+                    console.log("what");
+                    if (confirm('Logout?')) {
+                        firebase.auth().signOut(); // This will trigger onAuthStateChanged again, immediately.
+                    }
+                };
+            } else {
+                // if not logged in yet
+                window.location.href = 'index.html';
+
+            }
+        });
+    </script>
+
+
+</body>
+
+</html>
+
+
+
+
+
+
+
+
+
+
+
+
+
